Fix(as-types): moved Person serializable object to a proper dto test folder
