+++ packages/as-types/assembly/__tests__/serializer.spec.ts
@@ -1,4 +1,4 @@
-import { Divinity, Hero, Person } from '../Person';
+import { Divinity, Hero, Person } from './dto-tests/Person';
 import {
   boolToByte,
   bytesToF32,
